+++ pystac/__init__.py
@@ -18,9 +18,10 @@ from pystac.version import (
     get_stac_version,  # type:ignore
     set_stac_version,  # type:ignore
 )
+from pystac.media_type import MediaType  # type:ignore
+from pystac.rel_type import RelType  # type: ignore
 from pystac.stac_io import StacIO  # type:ignore
 from pystac.stac_object import STACObject, STACObjectType  # type:ignore
-from pystac.media_type import MediaType  # type:ignore
 from pystac.link import Link, HIERARCHICAL_LINKS  # type:ignore
 from pystac.catalog import Catalog, CatalogType  # type:ignore
 from pystac.collection import (