+++ tests/test_link.py
@@ -84,6 +84,19 @@ class LinkTest(unittest.TestCase):
 
 
 class StaticLinkTest(unittest.TestCase):
+    def setUp(self) -> None:
+        self.item = pystac.Item(
+            id="test-item",
+            geometry=None,
+            bbox=None,
+            datetime=TEST_DATETIME,
+            properties={},
+        )
+
+        self.collection = pystac.Collection(
+            "collection id", "desc", extent=ARBITRARY_EXTENT
+        )
+
     def test_from_dict_round_trip(self) -> None:
         test_cases: List[Dict[str, Any]] = [
             {"rel": "", "href": ""},  # Not valid, but works.
@@ -104,13 +117,21 @@ class StaticLinkTest(unittest.TestCase):
                 pystac.Link.from_dict(d)
 
     def test_collection(self) -> None:
-        c = pystac.Collection("collection id", "desc", extent=ARBITRARY_EXTENT)
-        link = pystac.Link.collection(c)
+        link = pystac.Link.collection(self.collection)
         expected = {"rel": "collection", "href": None, "type": "application/json"}
         self.assertEqual(expected, link.to_dict())
 
     def test_child(self) -> None:
-        c = pystac.Collection("collection id", "desc", extent=ARBITRARY_EXTENT)
-        link = pystac.Link.child(c)
+        link = pystac.Link.child(self.collection)
         expected = {"rel": "child", "href": None, "type": "application/json"}
+        self.assertEqual(expected, link.to_dict())
+
+    def test_canonical_item(self) -> None:
+        link = pystac.Link.canonical(self.item)
+        expected = {"rel": "canonical", "href": None, "type": "application/json"}
+        self.assertEqual(expected, link.to_dict())
+
+    def test_canonical_collection(self) -> None:
+        link = pystac.Link.canonical(self.collection)
+        expected = {"rel": "canonical", "href": None, "type": "application/json"}
         self.assertEqual(expected, link.to_dict())