+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ### Added
 
+- Added `Link.canonical` static method for creating links with "canonical" rel type ([#351](https://github.com/stac-utils/pystac/pull/351))
+- Added `RelType` enum containing common `rel` values ([#351](https://github.com/stac-utils/pystac/pull/351))
+
 ### Changed
 
 ### Fixed
@@ -26,7 +29,7 @@
 - Removed type information from docstrings, since it is redundant with function type
   annotations ([#342](https://github.com/stac-utils/pystac/pull/342))
 
-## [1.0.0-beta.1]
+## [v1.0.0-beta.1]
 
 ### Added